requirements/test.txt
django-debug-toolbar==5.1.0
parameterized==0.9.0
daphne==4.1.2
-r prod.txt